Changes the JDBC settings to only sepcify a fetchSize if more then 1000 rows will be fetched.
Other constellations would be quite pointless, as the fetch size is set to 1000. Especially
the AS400 JDBC driver thinks this is annoying enough to throw an exception.

Note that MySQL does not support a fetch size, which is why we set it into streaming mode.
MySQL is identified by the STREAMING capability here and Integer.MIN_VALUE is the
magic constant specified by the driver to perform streaming.