Adds a metric and limits for slow queries
The system logs each SQL query which takes longer than 10s (jdbc.logQueryThreshold).
A metric was aded to show the number of slow queries per minute and a limit was installed
so that the system state toggles to yellow if more than one slow query occurs per
minute (per default).