For relational operations like <, <=, >=, > we now default to "ignore null".
There is almost no sane standard or good reason to compare against null
in such cases. Using eq / ne / filled / notFilled is way clearer.

On the other hand there are a lot of occasions, where one wants
to filter using a given value if it is filled and ignore it otherwise.