Simplifies the logic used to detect is an EntityRef contains an Entity or ID.
We had several problems with methods overloading itself with
generic types. Most of the info required to identify the right
method to call is lost as runtime. Especially tagliatelle has
a hard time figuring out which method to invoke, as all
generics are lowered to their bound, therefore is(Object)
always wins the contest - which isn't quite right.