Apply date ranges correctly
The DateRange filter was not applied correctly as the gte/lte call does only return the constraint but does not apply it to the given query. Now the filter gets applied. The casting is needed as otherwise this would not be a valid Java call. Don't ask why ...

Tags: filter, date