Allows to generate bucket lists from JSON
The function getAggregationBuckets only allows to retrieve buckets from
the first level of aggregations. But because aggregations can be nested
we need the possibility to return the bucket lists from aggregations in
deeper levels. Also buckets can contain further aggregations so we need
to be able return the JSON as well.