Permits to specify a seconary database for OMA queries.
This can be used to query a local JDBC slave for "non-critical" queries.
This is beneficial in two way: First, the main database has reduced load
and second, the local slave might respond quicker as there is no
network overhead involved.

Note however, being a slave, the database might be out of sync,
therefore this must be only used to "uncritical" reads where a
delayed consistency (or even a transient inconsistency) is
acceptable.