Only apply an index defined by a superclass if none of the subclasses defines it.
For each entity we inherit all index definitions from all superclasses.
To permit a concrete entity type to overwrite the index defined by one
of its superclasses, we only apply the index if no other index with
the same name already exists.

As we iterate from bottom up (starting at the concrete class and walk
up to each superclass), this check is sufficient and gives priority
to the concrete implementations.