We need to set the closed connection to null in order to create a new one
If we do not reset the stmt variable we will run into an error after the connection was closed.
This might happen whenever an sqlException is thrown. The connection is closed but we are not able
to recover it because the object is not null and our build sql function requires a null variable to
create a new connection