Fixes MongoDateRangeFacets
When the Field-Value of the Document is not within the boundaries, it is counted to the default bucket. When there is not a single result, there is only the default bucket and "findFirst" returned the default bucket and in there the "count" was the number of all documents matching the query.

So all DateRanges without a single match were set to the total count of documents.