Removes a helper function which applied the accessPath too often (in some cases).
We cannot extract this functionality into a single method as sometimes we need to apply the
access path and sometimes we don't. Also, the catch block created an entirely wrong log message
and shouldn't tigger anyway.

Therefore the resorted to the original implementation and only utilize the accessPath if
appropriate.