Permit safeClose to be called when cascading errors from individual queries.
As after a SQL exception the database connection might be in an inconsistent state,
we close all open connections and try to re-open them.

However, the internal fields were nulled here to prevent a "use after free/close" bug.
This is now moved to the explicit close() call so that safeClose can be use to
"reset/restart" a BatchContext.