Removes sending the doc type in ES REST calls
This change is necessary as the new Elasticsearch version supports only one doc type per index.
We already did this in most cases but sent the type anyways but it has to be emitted for the new version as otherwise a multitude of deprecation warnings and exceptions are returned by the ES server.

OX-5554