We no longer report an error if an exception is encountered when re-setting / restoring the auto-commit flag.
If e.g. GaleryDB encounters an error and aborts a transaction the internal connection is in an inconsistent
state. If we now try to reset the auto-commit flag for the already dead connection a NPE like
"Connection is null" is reported. This error is misleading and therefore not logged at all as the previous error
is the actual root cause.