We now let to user control whether comma-separated strings are parsed into lists automatically.
Depending on the UI architecture (single entry field vs. multi-field) this is a very good or a
very bad idea. However, the user (programmer) should know which one to pick.

This is enable auto-parse when using an single text field which shows / accepts a
comma-separated string or leave it disabled if a multi-select field or the like is used.