Generates long (fully qualified foreign key names again).
An FK has to be GLOBALLY unique - not just within the table
(when using MariaDB). Therefore we add the table name and still
apply the limit to prevent the original problem of "too long FK names".

Note that the hard limit is 64 - so we stay a bit away from that by
limiting the result to 60 characters.