Reverts the change which shortened foreign key names.
We now generate the "classic" FKS names as long as these are
less than 60 chars long (64 is the hard limit). If a key name is too
long, we generate a shorter name and also limit this to 60 chars.

The prevents us from re-generating lots of FKs which is quite an
intense task.