Adds 'wasCreated' method into BaseEntity to be called after save
As the correct way to check this at this point is not obvious.
Without this the most obvious but faulty way was to use 'isNew()' here - but this would always return false as the id is already set at this point. The persistedData of the entity is not yet updated so we can check if the ID has been changed.
All provided tests would fail with 'isNew()' in the after save handler.