Code cleanup and refactoring.
We previously had a circular dependency where the LowLevel
client was the base for Elastic but itself referred to it. We now made
the LowLevel client a real low-level class which only encapsulates the
calls to ES and doesn't care about EntityDescriptors etc.

Also some common keys were extracted as constants.