Adds a helper method which can be used by schema evolution tools.
As mappings commonly cannot be updates in ES, we provide various
tools to support schema evaluation.

One tool is the newly introduced ability to split an index into a read
inex (using the old mappings) and a write index, using the new
mappings - when can be created, installed and committed during full
operation.