Permits to distinguish the routing suppressions for the read and the write index.
With the previous changes we introduced the possibility to redirect all writes for
and entity into a new index. We now also support to enable and disable the
routing suppression of both indices independent of each other.

This can be achieved using the "es-index" command. Using this facility one can
migrate from an unrouted index to a routed one an vice versa.