Uses MethodHandles in favor of reflection when invoking handlers.
With method handles, all access checks are performed when the
schema / EntityDescriptor is built. After that (when the handlers
are invoked a bazillion times) nothing will happen and the methods
are more or less directly invoked and also represented as a single
frame in a stack trace).

This should provide quite some performance improvements and
also increase the readability of stack traces.