Changes order of dropping foreign keys and keys / indices in the SchemaTool.
We previously first generated the DROP INDEX statements for all superfluous
key/indices and then handled foreign keys. This is a problem, at least with
MySQL/MariaDB, as every foreign key needs its matching index.

Therefore if a foreign key is superfluous, we first have to delete the constraint
and then the index. To achieve this, the order of the generated statements was
flipped.