Marks adding foreign keys as "possible dataloss".
Technically this isn't true - creating a foreign key constraint
should / must never lead to dataloss. However, we do not
want to create these constraints in an uncontrolled manner
if the underlying table wasn't new / is modified.

The reason for this is, that if we change the table which contains
the foreign key, we technically drop the old key, drop the
associated index and then create the new key. If we now
immediately create the new foreign key (as it is marked as
"no dataloss" and later on delete the old fk, we might end up
with the situation that the new fk uses the index which was created
for the old fk and therefore has the wrong name. Our tool would still
try to delete this index (without success, as it is still in use).

We therefore mark all three operations as "dataloss possible" so that
they are executed in proper order and in a controlled manner.