Properly detects if a column needs a change and tries to emit an appropriate SQL.
Note that changing columns is a very expensive task in clickhouse and should be
avoided if possible.

Also note that renaming columns will be implemented in Clickhouse v20.4 which has
not yet been released. We however emit the SQL, so that the system maintainer
knows that a column change is present - however, executing the statement will
fail.