Permits to mark queries as long running.
This suppresses some warnings about long running queries and
connections which are some times expected (e.g. when clickhouse
crunches billions of rows). Note that however, each query is still
wrapped in an operation and thus can be monitored. This will also
trigger an alarm if a query takes longer than 15 minutes.