Considers nullability before actually checking if property changed
This is necessary as the annotation `SkipDefaultValue` allows to omit storing the default values of properties in the database (which means in MongoDB the field is actually completely missing from the sorted object). When a field has any other default value than null isChanged would always return true before, as the equality check was comparing "null" (as the property is not in the persistedData map) against the default value of the entity (for example Amount.NOTHING).

We now utilize the `isConsideredNull` method of the property which effectively normalizes null and Amount.NOTHING to the same thing and compare them before running the equalsFunction.