Improve the iterate blockwise using row values
Background: The skip/limit construct is implemented rather slow in the database.
Because we know the value sorted right above the one we want to continue with, we can instead use this value, in conjunction with a total ordering to remove the entities already read from the result set.
This can result in a significant speedup, especially if you have the right indices, and even more so when there are expensive joins involved.