Allows to use a script source as a query score
Previously only a complex function score was possible (which most of the time means all query fields need to be scored as well to get a score other than 0). This basically implements https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-script-score-query.html which is a simple way of scoring the results of a query by a Plainless script. A common use case is "sorting" a query by random so we also provide a small helper function for that.