Warns if a SQL query references an unknown variable.
If a query like `SELECT * FROM table where col = ${test}` is executed
without providing a value for test, we log this warning, as this is
most probably a typo. If a value is unknown/unset, null should be used.

Note that this will only log this warning and still execute the query
as if "null" had been provided. Therefore this change is non-breaking.