BREAKING: Gives all primitive number properties a default value of 0
As we can't mark a primitive field as empty until the beforeSave handler checks for nullability, we can't produce proper error messages - so we always give them a default value of 0 so they behave just like in the java language.
Although this could also be considered a bugfix, this is marked as BREAKING as some functionality might rely on the IllegalArgumentException being thrown.
Developers should look at all primitive number fields and change them to non-primitive non-nullable if they want an error to be thrown/shown when empty values are given.

FixeS: OX-7315