BREAKING: Gives all properties a default value determined by their initial value
The biggest change from this is that all primitive number fields have a default value of '0' if not initialized with another number.
Previously they produced an IllegalArgumentException when they had no default value.
This means, for non-nullable fields that should not have a default value but should produce an error when left empty, a non-primitve datatype should be used. Then they will be properly transformed from import/webcontext to null and produce an error on save.

Please note that for primitive booleans this was already implemented, but if you used the default value directly, the actual default value has changed from '0' to 'false'.

Fixes: SIRI-415