BREAKING:Changes Property#defaultValue type from String to Value
Instead of a string that doubles as a value used for the column default value, the defaultValue will now hold an Object that represents the property wrapped in a Value, or Value.EMPTY if no default Value is present. JDBC Column default value can be retrieved via a new method.

To make it more universal, DefaultValue Annotations are now transformed like machine input (Property#transformValueFromImport), so existing DefaultValues might need to be adjusted, especially times,dates and booleans.

Fixes: SIRI-415