Enables to remove an entity from a cache by only specifying its id.
This is useful in cases where two entities are related but changing one does not directly affect a persisted field of the other, but rather a transient field. Instead of modifying the transient field, we may want to load a clean instance of the other entity if it is needed again.

- Fixes: OX-7421