Provides a "non-blocking" way of reindexing an entity.
Previously, we started an async request and awaited its completion.
However, now the internal http stack of the client will detect a stale
connection and kill it.

We therefore provide an API to start a re-index task and another method
to check if the task is still running. This way, we can manually poll
for completion.