Allows ElasticSearch delete by query to throw an OLE
In contrast to the believes when this method is introduced delete by query can actually result in an optimistic locking error when one of the entities that would have been deleted is modified after the deletion query is started.

The explanation in ES documentation is the following:
When a delete by query is executed a snapshot of the index is made and when the version of an entity that is being deleted differs from the one of the snapshot the deletion query is aborted and an OLE is reported.