Revises parsing of nested mappings 🗺
In the query tool, we use the syntax ABC.XYZ to parse nested properties of database entries. We used to require ABC to represent a reference to another entity holding a property XYZ. However, we may encounter cases (particularly in MIO) where we carry maps within an entity, and we want to be able to scan them as well. In order to do this, we now fall back to treating ABC.XYZ as a simple nested mapping when the assumption described above is not met. Prior to this commit, null would have been returned.