Fixes a sneaky injection of an Integer object into a Long field.
As Elasticsearch stores numbers as Integer, we might end up with
an SQLEntityRef which contains an Integer instead of a Long within
its id field. Under normal use, this will be autoboxed back and
forth by the JVM, but when calling some methods, strange class cast
or event invalid class exceptions start to occur.

We therefore convert the integer into a long.