Saves the generated id in elastic entities upon insert
The generated id was passed over as document id, but was not set as an id field in the actual document data, causing all new inserted entities to have a missing id.

The first update after it sets the id, because the doc id is set in the entity when reading, but the isChanged detects the id is missing.

Fixes: SIRI-594