add ORDERED to the characteristics
The stream is ordered (that means, it has a defined encounter order), since we need a defined ordering to do a block-wise iteration. We do not signal SORTED because we'd need to supply a comparator, which is currently too much effort for little gain.
Also, removing the qualification on the constants that we don't need here.