Warns about wrong nullable annotation on entity properties
Its a realistic error (that already happened in at least one case) that optional fields in an entity where annotated with Nullable instead of the mixing annotation NullAllowed.
If we detect a wrong annotation we output a warning at system startup.
As the intent by the developer that added the wrong is pretty clear we still act like the correct annotation was used in such cases to prevent wrong schemas from being created.