Provides a more robust way to query across multiple ES indices
Mainly this has the developer specify the common super class of the queried entity types and lets the compiler assert that common superclass is actually extended by all provided entities.

Note: it is still possible to do foolish things with this when e.g. specifying ElasticEntity as the supertype. Not a lot of useful things can be done with such queries.

Fixes: OX-10043