Allows to define shared validators for properties
This allows to define common validations that are used on fields across multiple entities.
Example could be something like email validation.
The alternative of wrapping single fields in composites can be kind of cumbersome and lead to ugly field names and fragmentation.

With the new `ValidatedBy` annotation a validator can be named that should be called in the validation and before-save phases of the property/entity lifecycle.
The validator has to be globally registered and implement the `PropertyValidator` interface while specifying a unique name.