BREAKING CHANGE: Fixes handling of empty lists when storing them in a SQL based DB without lists support (e.g. MariaDB).
The field is now nulled instead of filled with an empty string. Before 2021, empty lists were already stored as nulls, but this translation between an empty string to a null seems to have been removed.

This makes the FilterFactory#isEmptyList constraint work again correctly as it checks "IS NULL" and makes non null allowed StringList fields throw an error again for empty lists.

Breaking Change 1: StringList fields for SQL-based entities without list support (e.g. MariaDB, but not Clickhouse) need to be updated from an empty string to null.
Breaking Change 2: StringList fields without the NullAllowed annotation are now throwing an exception if an entity update tries to update the field with an empty list.

Fixes: SIRI-881