Add a metric for ES memory pressure
The memory pressure of Elasticsearch is supposed to be a good indicator for the performance of the cluster. It is calculated as the usage in percent of the old gen memory pool, as described in https://www.elastic.co/guide/en/elasticsearch/reference/current/high-jvm-memory-pressure.html

We also make the metric available over a public method. This can be used to slow down or pause time-uncritical tasks.

Fixes: SE-14021