Allow to control whether client info is sent for each redis pool
This was disabled globally in SIRI-921 as our Jupiter server seem to have problems with it.
As it turns out it answers the calls correctly but struggles with the way jedis sends 3 calls and expects 3 responses in bulk during the connection phase in Connection#initializeFromClientConfig.
To not waste much time on something that is actually ignored on the receiving side of jupiter anyway we just make it configurable to be able to enable it for the real redis connection and disable it for the jupiter connection.

Fixes: SIRI-971