Considers the timezone storing Dates in ClickHouse
The Date type has no timezone information. But the driver will use the provided date year/month/day to build an instance which can land in the previous day when the server's timezone is before the application timezone.

To trick this out we define a capability which can be defined to "trick" this out.

This has no effect during reads

Fixes: SIRI-1084