Stores dates using the UTC calendar
Clickhouse uses fromUnixTimestamp64Nano to pass the nanos from a date to the database. The nanos will be obtained after creating an Instant from the date using the local timezone Calendar. Since this function expects nanos in UTC, we must set it with a UTC Calendar to avoid date shifting.

Fixes: SIRI-1084